add support for test-local
