+++ test/api/responsetimer_test.js
@@ -17,8 +17,8 @@ suite("api/responsetimer", function() {
   });
 
   if (!cfg.get('influxdb:connectionString')) {
-    console.log("Skipping 'ResponseTimerTest', missing config file: " +
-                "taskcluster-base-test.conf.json");
+    throw new Error("Skipping 'ResponseTimerTest', missing config file: " +
+                    "taskcluster-base-test.conf.json");
     return;
   }
